Move LocationHelper to its own package
